🐛 Escape XML specials in the default NFT image title
The title rendered into the fallback SVG comes from the on-chain class
name, but only `&` was escaped. A name containing `<` produced malformed
XML that sharp refuses to parse, 500ing the class image endpoint.

Escape inside addTextOnImage so the SVG builder owns escaping its own
interpolation. Reuse ses.ts's identical escapeHtml by moving it to
misc.ts, which is side-effect free — importing it from ses.ts would pull
AWS SES client construction into the image request path.
